LayoutTests/webrtc/libwebrtc/release-while-setting-local-description-expected.txt

CONSOLE MESSAGE: line 1: Unhandled Promise Rejection: TypeError: Argument 1 ('candidate') to RTCPeerConnection.addIceCandidate must be an instance of RTCIceCandidate
CONSOLE MESSAGE: line 1: Unhandled Promise Rejection: InvalidStateError (DOM Exception 11): The object is in an invalid state.
PASS successfullyParsed is true

TEST COMPLETE
This test makes sure that RTCPeerConnection backend will free itself correctly even if released from the network thread.
